Change source structure to improve the component hierarchy
