+++ components/web/apps/org.wso2.carbon.apimgt.publisher/src/main/pages/src/app/components/Base/Footer/Footer.js
