+++ components/web/apps/org.wso2.carbon.apimgt.publisher/src/main/pages/{*}/{*}/index.html
@@ -1,0 +1,181 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <!-- The above 3 meta tags *must* come first in the head; any other head content must come *after* these tags -->
+    <title>WSO2 Theme Template</title>
+
+    <!-- Font WSO2 CSS -->
+    <link href="/publisher/public/components/root/base/javascripts/libs/font-wso2-1.2.1/css/font-wso2.min.css"
+          rel="stylesheet" type="text/css"/>
+
+    <!-- WSO2 Theme -->
+    <link href="/publisher/public/components/root/base/stylesheets/theme-wso2.min.css" rel="stylesheet">
+
+    <link href="/publisher/public/components/root/base/javascripts/libs/bootstrap-material-design/css/bootstrap-material-design.css"
+          rel="stylesheet">
+    <link href="/publisher/public/components/root/base/javascripts/libs/bootstrap-material-design/css/ripples.css"
+          rel="stylesheet">
+
+    <link rel="stylesheet" type="text/css" href="/publisher/public/components/root/base/stylesheets/test.css"/>
+    <link rel="stylesheet" type="text/css"
+          href="/publisher/public/components/root/base/stylesheets/bootstrap_buttons.css"/>
+    <link rel="stylesheet" type="text/css"
+          href="/publisher/public/components/root/base/stylesheets/progress-wizard.min.css">
+
+    <!--<link rel="stylesheet" type="text/css" href="/publisher/public/components/root/base/stylesheets/demo.css" />-->
+    <link rel="stylesheet" type="text/css" href="/publisher/public/components/root/base/stylesheets/icons.css"/>
+    <link rel="stylesheet" type="text/css" href="/publisher/public/components/root/base/stylesheets/component.css"/>
+
+    <script src="/publisher/public/components/root/base/javascripts/modernizr.custom.js"></script>
+
+    <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
+    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
+    <!--[if lt IE 9]>
+    <script src="https://oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js"></script>
+    <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
+    <![endif]-->
+</head>
+
+<body class="sticky-footer">
+
+
+<!-- page-content-wrapper -->
+<div class="page-content-wrapper" id="react-root">
+
+</div>
+<!-- /page-content-wrapper -->
+
+
+
+<script src="/publisher/public/components/root/base/dist/index.js" type="text/javascript"></script>
+<script src="/publisher/public/components/root/base/javascripts/modernizr.custom.js"></script>
+<!-- jQuery (necessary for Bootstrap and WSO2 Theme JavaScript plugins) -->
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
+<!-- Include all compiled plugins (below), or include individual files as needed -->
+<script src="/publisher/public/components/root/base/javascripts/bootstrap.min.js"></script>
+<script src="/publisher/public/components/root/base/javascripts/theme-wso2.min.js"></script>
+
+<script src="/publisher/public/components/root/base/javascripts/jquery.transit.min.js"></script>
+<script src="/publisher/public/components/root/base/javascripts/classie.js"></script>
+<script src="/publisher/public/components/root/base/javascripts/mlpushmenu.js"></script>
+
+<script>
+    new mlPushMenu(document.getElementById('mp-menu'), document.getElementById('nav-icon1'));
+</script>
+
+<script>
+    var button, parentLi, chGrid, swaggerFormContainer, posa, parentPosa, containerWidth, moveThismuch;
+
+    $(".test_button").on("click", morphButton);
+
+    $("#btn-close-step1").on("click", function () {
+        //debugger;
+        var closeBtn = $(this);
+        parentLi.removeClass("active");
+        swaggerFormContainer.addClass("hide");
+        $("#btn-close-step1").transition({opacity: '0', zIndex: '0'}, function () {
+            $("#btn-close-step1").hide();
+        });
+        $("#btn-next-step").transition({opacity: '0', zIndex: '0'}, function () {
+            $("#btn-next-step").hide();
+        });
+        button
+            .transition({
+                width: '200',
+                borderRadius: '50% 50% 50% 50%',
+                zIndex: '0',
+                x: '0'
+            }, 200, 'cubic-bezier(0.4, 0.0, 0.2, 1);');
+        //button.on("click",morphButton);
+        console.log(moveThismuch);
+    });
+
+
+    function morphButton() {
+        button = $(this);
+        parentLi = $(this).parent();
+        chGrid = $(".ch-grid");
+        swaggerFormContainer = $("#swagger-form-container");
+        posa = button.offset();
+        parentPosa = chGrid.offset();
+        containerWidth = chGrid.width();
+        moveThismuch = posa.left - parentPosa.left;
+
+        if (!parentLi.hasClass("active")) {
+            parentLi.addClass("active");
+            button
+                .transition({borderRadius: '0 0 0 0', zIndex: '1', delay: 300}, 'cubic-bezier(0.4, 0.0, 0.6, 1)')
+                .transition({x: -(moveThismuch - 20)}, 'cubic-bezier(0.4, 0.0, 0.2, 1);')
+                .transition({width: containerWidth}, 'cubic-bezier(0.4, 0.0, 0.2, 1);', function () {
+                    swaggerFormContainer.removeClass("hide");
+                    $("#btn-close-step1,#btn-next-step").show();
+                    $("#btn-close-step1").transition({
+                        opacity: '1',
+                        x: 90,
+                        y: -parentPosa.top - 60,
+                        zIndex: '2',
+                        delay: 100
+                    }, 'cubic-bezier(0.4, 0.0, 0.6, 1)');
+                    $("#btn-next-step").transition({
+                        opacity: '1',
+                        x: -30,
+                        y: -(parentPosa.top - 40),
+                        zIndex: '2',
+                        delay: 300
+                    }, 'cubic-bezier(0.4, 0.0, 0.6, 1)');
+                    //button.off("click",morphButton);
+                    //button.removeClass("ripple-effect");
+                    //debugger;
+                    //$('.ink', button).remove();
+                    //button.unbind('mouseenter').unbind('mouseleave');
+                });
+
+
+        }
+
+        console.log(moveThismuch);
+    }
+
+    $(".animate").bind('oanimationend animationend webkitAnimationEnd', function () {
+        console.log("fin");
+    });
+
+
+    (function ($) {
+        $(".ripple-effect").click(function (e) {
+            var rippler = $(this);
+
+            // create .ink element if it doesn't exist
+            if (rippler.find(".ink").length == 0) {
+                rippler.append("<span class='ink'></span>");
+            }
+
+            var ink = rippler.find(".ink");
+
+            // prevent quick double clicks
+            ink.removeClass("animate");
+
+            // set .ink diametr
+            if (!ink.height() && !ink.width()) {
+                var d = Math.max(rippler.outerWidth(), rippler.outerHeight());
+                ink.css({height: d, width: d});
+            }
+
+            // get click coordinates
+            var x = e.pageX - rippler.offset().left - ink.width() / 2;
+            var y = e.pageY - rippler.offset().top - ink.height() / 2;
+
+            // set .ink position and add class .animate
+            ink.css({
+                top: y + 'px',
+                left: x + 'px'
+            }).addClass("animate");
+        })
+    })(jQuery);
+</script>
+
+</body>
+</html>
